Most of the code
This is the main file of my snake game, it was a test to see if I could make the game fully on my own.
When I had managed to make the game, I decided to continue and try to make it into a proper game others could enjoy. 
It's for sale right now on steam. https://store.steampowered.com/app/2109530/Classic_Snake/